LayoutTests/platform/wincairo/imported/w3c/web-platform-tests/loading/preloader-css-import-no-semicolon.tentative-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

PASS Imported inline CSS with no semicolon is not blocked on pending CSS
